refactor : RacingCarGrid를 service layer로 분리
